requirements.txt
odfpy
openpyxl<2
xlrd
xlwt
